@@ -1,4 +1,5 @@
-// swift-tools-version:5.9
+// swift-tools-version:6.0
+
 import PackageDescription
 
 let package = Package(
